move LogSegment
